[TUI] Major Refactor of TUI Code (#268)
* Renamed `app.py`

* Refactored to `tui_logic.py` and `project_config.py`

* Refactored to `custom_widgets.py`

* Refactored `TabScreen` and `ProjectSelector`

* Fixed bug preventing tabs from loading correctly

* Pass mainwindow to tabs

* Renamed `tui_logic.py` to `utils.py` and refactored `ErrorScreen` to `modal_dialogs.py`

* Renamed `tab_screen.py` to `project_manager.py` and refactored

* Renamed `create_tab.py` to `create_folders.py`

* Renamed `project_configs` and `project_select.py`

* Rename `transfer_tab.py` to `transfer.py`

* Renamed `ProjectSelector` to `ProjectSelectorScreen`

* Changed `tui_app.py` back to `app.py`

* Added banners to help navigate `configs.py`

* Refactored `configs.py` into separate scripts for clarity
