+++ datashuttle/tui/tabs/create_folders.py
@@ -1,0 +1,82 @@
+from time import monotonic
+
+from textual.widgets import (
+    Button,
+    DirectoryTree,
+    Input,
+    Label,
+    TabPane,
+)
+
+from datashuttle.tui import custom_widgets
+
+
+class CreateFoldersTab(TabPane):
+    """
+    From this tab, the user can easily create new project files
+    formatted according to the NeuroBlueprint specification.
+    """
+
+    def __init__(self, mainwindow, project):
+        super(CreateFoldersTab, self).__init__(
+            "Create", id="tabscreen_create_tab"
+        )
+        self.mainwindow = mainwindow
+        self.project = project
+        self.prev_click_time = 0.0
+
+    def compose(self):
+        yield DirectoryTree(
+            self.project.cfg.data["local_path"],
+            id="tabscreen_directorytree",
+        )
+        yield Label("Subject(s)", id="tabscreen_subject_label")
+        yield Input(id="tabscreen_subject_input", placeholder="e.g. sub-001")
+        yield Label("Session(s)", id="tabscreen_session_label")
+        yield Input(id="tabscreen_session_input", placeholder="e.g. ses-001")
+        yield Label("Datatype(s)", id="tabscreen_datatype_label")
+        yield custom_widgets.DatatypeCheckboxes()
+        yield Button("Make Folders", id="tabscreen_make_folder_button")
+
+    def on_directory_tree_directory_selected(
+        self, event: DirectoryTree.DirectorySelected
+    ):
+        """
+        Upon double-clicking a directory within the directory-tree
+        widget, replace contents of the \'Subject\' and/or \'Session\'
+        input widgets, depending on the prefix of the directory selected.
+        Double-click time is set to the Windows default duration (500 ms).
+        """
+
+        click_time = monotonic()
+        if click_time - self.prev_click_time < 0.5:
+            if event.path.stem.startswith("sub-"):
+                self.query_one("#tabscreen_subject_input").value = str(
+                    event.path.stem
+                )
+            if event.path.stem.startswith("ses-"):
+                self.query_one("#tabscreen_session_input").value = str(
+                    event.path.stem
+                )
+        self.prev_click_time = click_time
+
+    def on_button_pressed(self, event: Button.Pressed):
+        """
+        Enables the Make Folders button to read out current input values
+        and use these to call project.make_folders().
+        """
+
+        if event.button.id == "tabscreen_make_folder_button":
+            sub_dir = self.query_one("#tabscreen_subject_input").value
+            ses_dir = self.query_one("#tabscreen_session_input").value
+
+            try:
+                self.project.make_folders(
+                    sub_names=sub_dir,
+                    ses_names=ses_dir,
+                    datatype=self.query_one("DatatypeCheckboxes").type_out,
+                )
+                self.query_one("#tabscreen_directorytree").reload()
+            except BaseException as e:
+                self.mainwindow.show_modal_error_dialog(str(e))
+                return
